change api to support multiple rules per symbol and use bevy engine to render the examples

+++ src/scope.rs
@@ -1,15 +1,15 @@
 #[derive(Clone, Default)]
 pub struct Scope {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
-    pub sx: f64,
-    pub sy: f64,
-    pub sz: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
+    pub sx: f32,
+    pub sy: f32,
+    pub sz: f32,
 }
 
 impl Scope {
-    pub fn new(x: f64, y: f64, z: f64, sx: f64, sy: f64, sz: f64) -> Self {
+    pub fn new(x: f32, y: f32, z: f32, sx: f32, sy: f32, sz: f32) -> Self {
         Self {
             x,
             y,
@@ -19,21 +19,21 @@ impl Scope {
             sz,
         }
     }
-    pub fn set_size(&mut self, sx: f64, sy: f64, sz: f64) -> &Self {
+    pub fn set_size(&mut self, sx: f32, sy: f32, sz: f32) -> &Self {
         self.sx = sx;
         self.sy = sy;
         self.sz = sz;
 
         return self;
     }
-    pub fn extrude(&mut self, sy: f64) -> &Self {
+    pub fn extrude(&mut self, sy: f32) -> &Self {
         self.sy = sy;
 
         self
     }
     pub fn get_face(&self, face: Face) -> Self {
         match face {
-            Face::Top => Scope::new(self.x, self.y + self.sy, self.z, self.sx, 0f64, self.sz),
+            Face::Top => Scope::new(self.x, self.y + self.sy, self.z, self.sx, 0f32, self.sz),
             Face::Bottom => todo!(),
             Face::Right => todo!(),
             Face::Left => todo!(),
@@ -41,7 +41,7 @@ impl Scope {
             Face::Back => todo!(),
         }
     }
-    pub fn repeat(&self, dir: Direction, size: f64) -> Vec<Self> {
+    pub fn repeat(&self, dir: Direction, size: f32) -> Vec<Self> {
         match dir {
             Direction::X => {
                 let width = (self.sx / size).floor();
@@ -50,7 +50,7 @@ impl Scope {
                 let mut scopes = Vec::<Self>::with_capacity(10);
                 for x in 0..count {
                     let new_scope = Self::new(
-                        self.x + width * x as f64,
+                        self.x + width * x as f32,
                         self.y,
                         self.z,
                         width,
@@ -70,7 +70,7 @@ impl Scope {
                 for i in 0..count {
                     let new_scope = Self::new(
                         self.x,
-                        self.y + height * i as f64,
+                        self.y + height * i as f32,
                         self.z,
                         self.sx,
                         height,
@@ -90,7 +90,7 @@ impl Scope {
                     let new_scope = Self::new(
                         self.x,
                         self.y,
-                        self.z + breadth * i as f64,
+                        self.z + breadth * i as f32,
                         self.sx,
                         self.sy,
                         breadth,
